+++ R/ggpubr-package.R
@@ -6,7 +6,7 @@
 #'
 #' \itemize{
 #' \item \href{https://rpkgs.datanovia.com/ggpubr/}{ggpubr documentation}
-#' \item \href{https://www.datanovia.com/en/blog/tag/ggpubr/}{ggpubr tutorials}
+#' \item \href{https://www.datanovia.com/learn/data-visualization/ggpubr/}{ggpubr tutorials}
 #' }
 #'
 #' @section P-Value Formatting: